Move legacy chime-server into codegen sub-directory
