+++ src/app/clusters/chime-server/app_config_dependent_sources.cmake
@@ -16,6 +16,7 @@
 TARGET_SOURCES(
   ${APP_TARGET}
   PRIVATE
-    "${CLUSTER_DIR}/chime-server.cpp"
+    "${CLUSTER_DIR}/codegen/chime-server.cpp"
+    "${CLUSTER_DIR}/codegen/chime-server.h"
     "${CLUSTER_DIR}/chime-server.h"
 )
