+++ src/app/clusters/chime-server/app_config_dependent_sources.gni
@@ -12,6 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 app_config_dependent_sources = [
-  "chime-server.cpp",
+  "codegen/chime-server.cpp",
+  "codegen/chime-server.h",
   "chime-server.h",
 ]
